refact: GameManager 클래스를 생성하여 Game 클래스의 역할 분배
